Refactor harvester source classes a bit
