+++ pta-intelligent-search-harvester/src/main/java/fi/maanmittauslaitos/pta/search/source/Harvestable.java
@@ -1,4 +1,4 @@
-package fi.maanmittauslaitos.pta.search.csw;
+package fi.maanmittauslaitos.pta.search.source;
 
 public interface Harvestable {
 	String getIdentifier();
