+++ pta-intelligent-search-harvester/src/main/java/fi/maanmittauslaitos/pta/search/source/csw/CSWHarvestable.java
@@ -1,4 +1,6 @@
-package fi.maanmittauslaitos.pta.search.csw;
+package fi.maanmittauslaitos.pta.search.source.csw;
+
+import fi.maanmittauslaitos.pta.search.source.Harvestable;
 
 public class CSWHarvestable implements Harvestable {
